+++ CHANGELOG.md
@@ -38,3 +38,4 @@ Change log for the codebase. Initialised from the developments following version
     - shifted repeated fit code from TabAuto children into TabAuto
     - shifted repeated init code from TabAuto children into TabAuto
     - generalised repeated _define_model code and shifted from TabAuto children into TabAuto
+    - removed repeated set_params from TabAuto